refactor: heavy refactory of pages directory
