+++ src/app/pages/teacher/components/analytics/components/shortcut-card-placeholder/shortcut-card-placeholder.component.html
