+++ src/app/pages/teacher/components/analytics/components/shortcut-card/shortcut-card.component.html
