+++ src/app/pages/teacher/components/analytics/interfaces/shortcut-card.interface.ts
