+++ src/app/pages/teacher/components/navbar/navbar.component.scss
